3.3-permissions/api_with_restrictions/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-debug-toolbar==4.2.0
django-filter==23.3
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
